+++ src/hardis-plugins-provider.ts
@@ -10,10 +10,12 @@ import {
   RECOMMENDED_SFDX_CLI_VERSION,
   RECOMMENDED_MINIMAL_SFDX_HARDIS_VERSION,
   resetCache,
+  execCommandWithProgress,
 } from "./utils";
 import { Logger } from "./logger";
 import which from "which";
 import { ThemeUtils } from "./themeUtils";
+import { SetupHelper } from "./utils/setupUtils";
 
 let nodeInstallOk = false;
 let gitInstallOk = false;
@@ -490,6 +492,29 @@ export class HardisPluginsProvider
     await Promise.allSettled(pluginPromises);
     // Propose user to upgrade if necessary
     if (outdated.length > 0) {
+      const command = this.buildUpgradeCommand(
+        outdated,
+        plugins,
+        legacySfdx,
+        sfdxCliOutdated,
+      );
+      const setupHelper = SetupHelper.getInstance();
+      const config = vscode.workspace.getConfiguration("vsCodeSfdxHardis");
+      if (config.get("autoUpdateDependencies") === true && !setupHelper.hasUpdatesInProgress()) {
+        setupHelper.setUpdateInProgress(true, command);
+        execCommandWithProgress(
+            command,
+            { fail: false, output: true },
+            `Automatically upgrading dependencies with command: ${command}`,
+        ).then(() => {
+          setupHelper.setUpdateInProgress(false, command);
+          vscode.commands.executeCommand("vscode-sfdx-hardis.refreshPluginsView");
+        }).catch(() => {
+          setupHelper.setUpdateInProgress(false, command);
+          vscode.commands.executeCommand("vscode-sfdx-hardis.refreshPluginsView");
+        });
+      }
+      else {
       vscode.window
         .showWarningMessage(
           "🦙 Some plugins are not up to date, please click to upgrade, then wait for the process to be completed before performing actions",
@@ -497,27 +522,33 @@ export class HardisPluginsProvider
         )
         .then((selection) => {
           if (selection === "Upgrade plugins") {
-            let command = outdated
-              .map((plugin) => `echo y|sf plugins:install ${plugin.name}`)
-              .join(" && ");
-            if (legacySfdx) {
-              command =
-                "npm uninstall sfdx-cli --global && npm install @salesforce/cli --global && " +
-                plugins
-                  .map((plugin) => `echo y|sf plugins:install ${plugin.name}`)
-                  .join(" && ");
-            } else if (sfdxCliOutdated === true) {
-              command = "npm install @salesforce/cli -g && " + command;
-            }
-            command = command + ` && sf hardis:work:ws --event refreshPlugins`;
+
             vscode.commands.executeCommand(
               "vscode-sfdx-hardis.execute-command",
               command,
             );
           }
         });
+      }
     }
     return items.sort((a: any, b: any) => (a.label > b.label ? 1 : -1));
+  }
+
+  private buildUpgradeCommand(outdated: any[], plugins: any,legacySfdx: boolean, sfdxCliOutdated: boolean): string {
+    let command = outdated
+      .map((plugin) => `echo y|sf plugins:install ${plugin.name}`)
+      .join(" && ");
+    if (legacySfdx) {
+      command =
+        "npm uninstall sfdx-cli --global && npm install @salesforce/cli --global && " +
+        plugins
+          .map((plugin: any) => `echo y|sf plugins:install ${plugin.name}`)
+          .join(" && ");
+    } else if (sfdxCliOutdated === true) {
+      command = "npm install @salesforce/cli -g && " + command;
+    }
+    command = command + ` && sf hardis:work:ws --event refreshPlugins`;
+    return command;
   }
 
   // Compare two semver strings. Returns -1 if a < b, 0 if equal, 1 if a > b